fix(moe): preserve input dtype through expert routing (#45)
* fix(moe): preserve input dtype through expert routing

The routing one-hot and accumulator buffers were created as default-fp32
zeros, silently promoting the whole MoELayer output to fp32 for fp16
models (found via smeltr profiling of Hunyuan3D-2.1-mlx Stage 1: ~45% of
MoE gemm time ran in fp32 with per-use weight upcast copies). The PT
reference (Hunyuan3D moe_layers.py) casts back to hidden_states.dtype.


* chore(ci): pin ty to 0.0.52 — newer releases fail mlx.core resolution

Unpinned ty picked up a release that can no longer resolve mlx.core on
the CI runners (174 unresolved-import diagnostics) and errors on
now-unused ignore directives. 0.0.52 is the version the last green
main run used.


* chore(ci): constrain dev env to mlx<0.32 — 0.32.0 wheel dropped type stubs

Correction to the previous commit's diagnosis: ty 0.0.52 was not the
regression — mlx 0.32.0 (released since the last green run on 0.31.2)
no longer bundles mlx/core/*.pyi, so ty cannot resolve mlx.core in any
version. Verified locally: same ty against mlx 0.31.2 passes, against
0.32.0 fails with the same 77 diagnostics as CI. Runtime imports are
unaffected; this constrains only the dev/type-check environment.


* chore: retrigger ci


---------

+++ src/mlx_arsenal/moe/moe.py
@@ -91,12 +91,14 @@ class MoELayer(nn.Module):
         # expert_weights: (num_tokens, num_experts)
         num_experts = len(self.experts)
         num_tokens = hidden.shape[0]
-        expert_weights = mx.zeros((num_tokens, num_experts))
+        # Match the input dtype: default-fp32 zeros would silently promote
+        # the whole layer output to fp32 for fp16/bf16 models.
+        expert_weights = mx.zeros((num_tokens, num_experts), dtype=hidden.dtype)
         for k in range(indices.shape[1]):
             col_indices = indices[:, k : k + 1]  # (num_tokens, 1)
             col_weights = weights[:, k : k + 1]  # (num_tokens, 1)
             # One-hot for this top-k slot
-            one_hot = mx.zeros((num_tokens, num_experts))
+            one_hot = mx.zeros((num_tokens, num_experts), dtype=hidden.dtype)
             rows = mx.arange(num_tokens).reshape(-1, 1)
             one_hot[rows, col_indices] = col_weights
             expert_weights = expert_weights + one_hot
